Add files via upload
This model helps in finding the Breakeven Point for the number Ticket sales to start making profit for upcoming Garage Band. In this model, User can set the Target profit and model can suggest the minimum required tickets to be sold to meet the target. This model is generic in nature, This model can be used for different size of the capacity of the venue, varying ticket price, fixed costs, revenue based on sales number, and security personal required and their cost. Tools user are macro, VLOOKUP , Goal seek, complex functions and excel charts